backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.28.0
sqlmodel==0.0.14

# Kite Connect
kiteconnect>=4.2.0
websocket-client>=1.6.1  # Required for WebSocket support

# Data processing
pandas==2.1.3
numpy==1.26.2
ta==0.10.2

# Async
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Security
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6

# Utils
python-dateutil==2.8.2
pytz==2023.3.post1

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.0
flake8==6.1.0
pre-commit==3.5.0

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.1
mkdocstrings[python]==0.23.0